52 view to display all cards returned by the api

+++ app/src/main/java/com/github/sdp/tarjetakuna/model/MagicCard.kt
@@ -27,7 +27,7 @@ data class MagicCard(
      * The total card mana cost.
      * Example : 7
      */
-    val convertedManaCost: Int = 0,
+    val convertedManaCost: Double = 0.0,
 
     /**
      * The card mana cost.
